pyproject.toml
[tool.poetry]
name = "statcandb"
version = "0.1.0"
description = ""
authors = ["Kevin H Wilson <khwilson@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "statcandb", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.6"
requests = "^2.31.0"
pyarrow = "^12.0.1"
duckdb = "^0.8.1"
python-dotenv = "^1.0.0"
boto3 = "^1.28.57"
tqdm = "^4.66.1"
sqlalchemy = "^2.0.21"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
statcandb = "statcandb.cli.main:cli"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
pytest-httpserver = "^1.0.8"
pytest = "^7.4.0"
ruff = "^0.0.282"
pandas = "^2.0.3"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "I"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
